+++ test/controllers/MachinesAvailableControllerSpec.scala
@@ -41,7 +41,7 @@ class MachinesAvailableControllerSpec extends SpecBase with MockitoSugar {
 
   def onwardRoute = Call("GET", "/foo")
 
-  val validAnswer: Int = 10
+  val validAnswer: Long = 10
 
   val selectedReturn: SelectedReturn = SelectedReturn(LocalDate.of(2025, 1, 1), LocalDate.of(2025, 3, 31))
